Added add and remove methods for users on company controller and created companyservice to handle company business logic

+++ src/main/java/no/ntnu/stud/idata2306_project/repository/CompanyRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.ListCrudRepository;
 
 import no.ntnu.stud.idata2306_project.model.company.Company;
 
-public interface CompanyRepository extends ListCrudRepository<Company, Integer> {
+public interface CompanyRepository extends ListCrudRepository<Company, Long> {
 }
